+++ packages/theme/CHANGELOG.md
@@ -7,3 +7,4 @@ All notable changes to [@bpmn-io/theme](https://github.com/bpmn-io/themes) are d
 ___Note:__ Yet to be released changes appear here._
 
 * `FEAT`: add the bpmn.io design tokens
+* `FEAT`: add `bpmn-io-theme-validate` and `bpmn-io-theme-audit` to check the tokens a library copied